Move error handling to dedicated module
This will be used to also host the runtime error function.
