+++ src/uu/sed/src/delimited_parser.rs
@@ -8,28 +8,12 @@
 // For the full copyright and license information, please view the LICENSE
 // file that was distributed with this source code.
 
+use crate::error_handling::compilation_error;
 use crate::script_char_provider::ScriptCharProvider;
 use crate::script_line_provider::ScriptLineProvider;
-use std::char;
-use uucore::error::{UResult, USimpleError};
 
-// Fail with msg as a compile error at the current location
-pub fn compilation_error<T>(
-    lines: &ScriptLineProvider,
-    line: &ScriptCharProvider,
-    msg: impl ToString,
-) -> UResult<T> {
-    Err(USimpleError::new(
-        1,
-        format!(
-            "{}:{}:{}: error: {}",
-            lines.get_input_name(),
-            lines.get_line_number(),
-            line.get_pos() + 1,
-            msg.to_string()
-        ),
-    ))
-}
+use std::char;
+use uucore::error::UResult;
 
 /// Return true if c is a valid octal digit
 fn is_ascii_octal_digit(c: char) -> bool {
